+++ pkgin.h
@@ -238,7 +238,7 @@ Sumfile		*sum_open(char *, time_t *);
 int		sum_start(struct archive *, void *);
 ssize_t		sum_read(struct archive *, void *, const void **);
 int		sum_close(struct archive *, void *);
-ssize_t		download_pkg(char *, FILE *);
+off_t		download_pkg(char *, FILE *);
 /* summary.c */
 int		update_db(int, char **, int);
 void		split_repos(void);
